backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
supabase
scikit-learn
numpy
python-dotenv
